backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
sqlalchemy==2.0.23
psycopg2-binary
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==12.1.1
whitenoise==6.4.0
qrcode==7.4.2
zxing-cpp==2.3.0; python_version < "3.14"
